requirements.txt
appdirs==1.4.4
asgiref==3.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
cssselect==1.2.0
Django==5.1.1
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
fake-useragent==1.5.1
feedparser==6.0.11
idna==3.10
importlib_metadata==8.5.0
lxml==5.3.0
numpy==2.1.1
pandas==2.2.3
parse==1.20.2
pillow==10.4.0
psycopg2==2.9.9
pyee==11.1.1
pyppeteer==2.0.0
pyquery==2.0.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
requests-html==0.10.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
w3lib==2.2.1
websockets==10.4
yahoo-fin==0.8.9.1
zipp==3.20.2
